+++ Public/New-IntuneWin32AppDetectionRuleScript.ps1
@@ -19,10 +19,11 @@ function New-IntuneWin32AppDetectionRuleScript {
         Author:      Nickolaj Andersen
         Contact:     @NickolajA
         Created:     2020-09-17
-        Updated:     2020-09-17
+        Updated:     2021-08-31
 
         Version history:
         1.0.0 - (2020-09-17) Function created
+        1.0.1 - (2021-08-31) Fixed an issue when using a non-UTF encoded multi-line script file
     #>
     [CmdletBinding(SupportsShouldProcess = $true)]
     param(
@@ -45,7 +46,7 @@ function New-IntuneWin32AppDetectionRuleScript {
         # Detect if passed script file exists
         if (Test-Path -Path $ScriptFile) {
             # Convert script file contents to base64 string
-            $ScriptContent = [System.Convert]::ToBase64String([System.IO.File]::ReadAllBytes("$($ScriptFile)"))
+            $ScriptContent = [System.Convert]::ToBase64String([System.Text.Encoding]::UTF8.GetBytes([System.IO.File]::ReadAllBytes("$($ScriptFile)") -join [Environment]::NewLine))
 
             # Construct detection rule ordered table
             $DetectionRule = [ordered]@{