+++ Public/New-IntuneWin32AppRequirementRuleScript.ps1
@@ -76,10 +76,11 @@ function New-IntuneWin32AppRequirementRuleScript {
         Author:      Nickolaj Andersen
         Contact:     @NickolajA
         Created:     2020-04-29
-        Updated:     2020-04-29
+        Updated:     2020-08-31
 
         Version history:
         1.0.0 - (2020-04-29) Function created
+        1.0.1 - (2021-08-31) Fixed an issue when using a non-UTF encoded multi-line script file
     #>
     [CmdletBinding(SupportsShouldProcess = $true)]
     param(
@@ -205,7 +206,7 @@ function New-IntuneWin32AppRequirementRuleScript {
             $ScriptFileName = [System.IO.Path]::GetFileName("$($ScriptFile)")
 
             # Convert script file contents to base64 string
-            $ScriptContent = [System.Convert]::ToBase64String([System.IO.File]::ReadAllBytes("$($ScriptFile)"))
+            $ScriptContent = [System.Convert]::ToBase64String([System.Text.Encoding]::UTF8.GetBytes([System.IO.File]::ReadAllBytes("$($ScriptFile)") -join [Environment]::NewLine))
 
             switch ($PSCmdlet.ParameterSetName) {
                 "String" {